RFCT: refactor api call to own file and use constants for query, mutate and status
